Interface for LookupContext
To provide the ability to fake the Filter it is usefull to have an interface for the LookupContext instead of a class.
